Fixed deadlock in hid-gfb
The root cause of the deadlock seems to be one or all of the following:

1. Work queues implementing deferred IO do not allow spinlocks to be
locked when exiting the deferred task. This means the lock protecting
fb_vbitmap (left for the urb completion handler to unlock) was reported
as problematic by the lockdeps analyzer.

The lock was converted to a soft-lock (a boolean flag), which in turn is
locally protected by the original spinlock.

2. URB completion handler was run from an IRQ handler, which meant the
URB spinlock was used both in IRQ context and in process context. Again,
lockdeps complained so the spinlock is now used with the
irqsave/irqrestore version.